requirements.txt
edgartools==3.3.2
fastapi[standard]==0.115.6
mypy==1.14.0
pymongo==4.10.1
ruff==0.8.4
torch==2.5.1
transformers==4.47.1
uvicorn==0.34.0
